requirements.txt
tensorflow>=2.13.0
numpy>=1.23.5
scikit-learn>=1.2.2
flask>=2.2.3
pillow>=9.5.0
waitress>=2.1.2
matplotlib>=3.7.0 